requirements_inference.txt
datasets
hydra-core
omegaconf
onnxruntime
fastapi
uvicorn
wandb
